Move on_demand_[cadip|prip]_staging in their own module like auxip
